Stock-ms

+++ composition-pedido/src/main/java/org/example/compositionpedido/dto/ProductoDetalleDto.java
@@ -2,11 +2,13 @@ package org.example.compositionpedido.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductoDetalleDto {
     private Long productoId;
     private String nombre;
-    private Double precioUnitario;
+    private BigDecimal precioUnitario;
     private Integer cantidad;
-    private Double subtotal;
+    private BigDecimal subtotal;
 }
