+++ composition-pedido/src/main/java/org/example/compositionpedido/dto/ProductoDto.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductoDto {
     private Long id;
@@ -11,7 +13,7 @@ public class ProductoDto {
     private String nombre;
 
     @NotNull(message = "El precio es obligatorio")
-    private Double precio;
+    private BigDecimal precio;
 
     private Long categoriaId;
     private String categoriaNombre;
